feature: read query file from given path
Adds the ability to read queries from a given path. When the "exec"
command is issued, the string is considered a path and a "stat" is
performed against that potential path.

If it happens to be a file, then we attempt to read the contents.
Otherwise it is assumed to be a regular query that the user passed.