feature: remove almost all whitespaces from the queries
The given queries may contain leading and trailing white spaces, as well
as many white spaces in between. The goal of this change is to remove them
from the given query.

It also marshals the query with the JSON marshaller, so that any
specified double quotes in the query are escaped for us.