refactor: only allow ".sql" files to be read
Prior to this change, if an invalid file path was given to the "exec"
command, the given path was interpreted as the query to be run with
Gabi, which resulted in the database returning errors since it didn't
understand the statement.

In order to avoid that, we are restricting the files we read to those
which only end with an ".sql" extension. That way, we ensure that we
give feedback to the user about the invalid specified file, and we avoid
sending a bogus query to the Gabi instances.